Move reflection read to queries/, rename use_cases/ to commands/
The reflection module has one read (get) and one command (upsert). The read
is the trivial shape the guide describes — one ownership check plus one
find_by_book_id — so it takes the ADR's halfway option: it relocates to
queries/ as-is, still delegating to BookReflectionRepository and still
returning the BookReflection entity. No query port, no adapter, no DTOs.

With use_cases/ down to the upsert command, the package renames to commands/
per the ADR's end-state naming. The queries-are-dead-ends contract already
covers both package names since the tagging migration.

Also fix the guide's dead-end inspection grep, which globbed only
use_cases/ and so stopped covering modules once they were renamed.
