+++ backend/src/infrastructure/reflection/routers/book_reflections.py
@@ -5,11 +5,11 @@ from typing import Annotated
 from fastapi import APIRouter, Depends
 from starlette import status
 
-from src.application.reflection.use_cases.get_book_reflection_use_case import (
-    GetBookReflectionUseCase,
-)
-from src.application.reflection.use_cases.upsert_book_reflection_use_case import (
+from src.application.reflection.commands.upsert_book_reflection_use_case import (
     UpsertBookReflectionUseCase,
+)
+from src.application.reflection.queries.get_book_reflection_use_case import (
+    GetBookReflectionUseCase,
 )
 from src.core import container
 from src.domain.identity import User
